pyproject.toml
# Copyright (c) 2025 Joan Fabrégat <j@fabreg.at>
# Permission is hereby granted, free of charge, to any person
# obtaining a copy of this software and associated documentation
# files (the "Software"), to deal in the Software without
# restriction, subject to the conditions in the full MIT License.
# The Software is provided "as is", without warranty of any kind.

[project]
name = "document-processor"
version = "v0.2.1"
description = "A document processor based on Docling and Tesseract."
authors = [
    { name = "Joan Fabrégat", email = "jfabregat@fogandfrog.com" }
]
requires-python = ">=3.12"
dependencies = [
    "docling>=2.28.4",
    "fastapi>=0.115.11",
    "pydantic>=2.11.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tesserocr>=2.8.0",
    "python-dotenv>=1.1.0",
    "pymupdf>=1.25.5",
]

# https://docs.astral.sh/uv/guides/integration/pytorch/#using-a-pytorch-index
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]
docling = { git = "https://github.com/joanfabregat/docling.git", rev = "fix-1333" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
